% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_list}
\alias{collapse_list}
\title{Collapse studies in a list}
\usage{
collapse_list(expr_list)
}
\arguments{
\item{expr_list}{A list with expression data elements}
}
\value{
A list with collapsed expression data
}
\description{
This function takes a list of expression data and applies the collapse_rows function
}
