% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_rows}
\alias{collapse_rows}
\title{Collapse redundant probes into gene symbols}
\usage{
collapse_rows(expr, probe_col, gene_col, data_table = FALSE,
  method = c("max_mean", "min_mean", "col_mean", "col_median"))
}
\arguments{
\item{expr}{Object of class \code{data.frame} containing expression values,
a column with probe symbols and a column with gene symbols}

\item{probe_col}{Character string specifying the name of the column containing 
probe names. Can also take rownames as input using the string "rownames".}

\item{gene_col}{Character string specifying the name of the column containing 
gene symbols to be used to collapse.}

\item{data_table}{Logical. If \code{TRUE} will return a \code{data.table} object.
Defaults to \code{FALSE}.}

\item{method}{Method used to collapse the probes. Can take one of "max_mean", 
"min_mean", "col_mean" or "col_median".}
}
\value{
Object of class \code{data.frame} or \code{data.table}
}
\description{
This function collapses redundant probe expression rows into single rows
based on repeated gene symbols in a given column.
}
