% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{get_annot_from_gse}
\alias{get_annot_from_gse}
\title{Get annotation data from GEO using a GSE object and
create a tentative sample_annotation file.}
\usage{
get_annot_from_gse(gse)
}
\arguments{
\item{gse}{The GSE object for the study}
}
\description{
The function tries to see which 
columns in the data provide information about each sample.
}
