% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{get_expr_from_gds}
\alias{get_expr_from_gds}
\title{Get expression data from a GDS object}
\usage{
get_expr_from_gds(gds, do_log2 = TRUE)
}
\arguments{
\item{gds}{The GDS object}

\item{do_log2}{Whether or not to apply log2 transformation on the data
(Default: TRUE)}
}
\description{
This function is used to extract expression data of a 
GDS object as given by get_gds.
}
