% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{get_expr_from_gse}
\alias{get_expr_from_gse}
\title{Get expression data from GEO using the study's GSE accession number}
\usage{
get_expr_from_gse(gse_id)
}
\arguments{
\item{gse_id}{The GEO accession ID for the study}
}
\description{
This function is used to query GEO and extract expression data of a 
study using the accession number as ID.
}
