% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{get_gpl_related_gds}
\alias{get_gpl_related_gds}
\title{Get GDS ids whose related studies are in the given platform}
\usage{
get_gpl_related_gds(gpl_id, min_sample_num = NULL,
  max_sample_num = NULL, verbose = FALSE)
}
\arguments{
\item{gpl_id}{The ID of the platform of interest}
}
\description{
Get GDS ids whose related studies are in the given platform
}
