% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{get_gds}
\alias{get_gds}
\alias{get_gse}
\title{Get GSE object from GEO using the study GSE accession number}
\usage{
get_gds(gds_id)

get_gse(gse_id)
}
\arguments{
\item{gse_id}{The GEO accession ID for the study}

\item{gse_id}{The GEO accession ID for the study}
}
\description{
This function is a wrapper function for getGEO, in which it will 
download GDS data.

This function is a wrapper function for getGEO, in which it will 
download the GSE matrix and get gpl data.
}
