% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{get_sample_number_studies}
\alias{get_sample_number_studies}
\title{Get studies with a specified range of samples}
\usage{
get_sample_number_studies(expr_list, min_samples, max_samples)
}
\arguments{
\item{expr_list}{A list object in which each element is an expression data.frame with 
a column with gene symbols and probe names in the rownames}

\item{min_samples}{The minimum number of samples to be included (inclusive)}

\item{max_samples}{The maximum number of samples to be included (inclusive)}
}
\description{
This function filters a list of expression data and returns the list 
without studies which fall outside of the specified sample number range.
It assumes that there is a column with gene symbols in each expression data, 
so it subtracts 1 from the number of columns.
}
