% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{insert_sample_info}
\alias{insert_sample_info}
\title{Insert sample information}
\usage{
insert_sample_info(pool, annot, study_id)
}
\arguments{
\item{pool}{A connection object}

\item{annot}{A data frame containing sample annotation information}

\item{study_id}{ID of the study in the study_info table}
}
\description{
This function takes information about samples and adds it into the 
sample_annot table in the database.
}
