% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{run_cemitool}
\alias{run_cemitool}
\title{Run CEMiTool}
\usage{
run_cemitool(expression, output_dir, study_id, annotation, interactions,
  gmt_fname)
}
\arguments{
\item{expression}{The name of the expression file to be analyzed}

\item{output_dir}{The name of the folder to output the results}

\item{study_id}{The GEO accession number (GSE) of the study}

\item{annotation}{The name of the sample annotation file}

\item{interactions}{The name of the interactions file}

\item{gmt_fname}{The name of the gmt file}
}
\description{
This function is used to run the CEMiTool.R script included in the CEMiTool package.
}
