% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{write_expr_data}
\alias{write_expr_data}
\title{Get expression data tables}
\usage{
write_expr_data(gse_ids, min_samples = 15, max_samples = 200,
  directory = file.path(getwd(), "data"), force = FALSE,
  force_repeat = FALSE)
}
\arguments{
\item{gse_ids}{A character vector containing GEO study accession IDs}

\item{min_samples}{The minimum number of samples for studies (inclusive)}

\item{max_samples}{The maximum number of samples for studies (inclusive)}

\item{directory}{The directory to save files}

\item{force}{If TRUE, overwrites the given directory.}

\item{force_repeat}{If FALSE, avoids downloading study data already present
in the 'directory' folder (as informed by folder names).}
}
\description{
This function takes a vector of GEO study IDs, selects those with only one platform and 
those with samples within a given range. It then uses the collapse_rows function to collapse
probes, and finally writes the expression data to files in the given directory.
}
