% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{get_expr_from_gse}
\alias{get_expr_from_gse}
\title{Get expression data from a GSE object}
\usage{
get_expr_from_gse(gse)
}
\arguments{
\item{gse}{The GSE object}
}
\description{
This function is used to extract expression data of a 
GSE object as given by get_gse.
}
