% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{get_platform_gse_ids}
\alias{get_platform_gse_ids}
\title{Get platform GSE ids}
\usage{
get_platform_gse_ids(gpl_id)
}
\arguments{
\item{gpl_id}{The ID of the platform of interest}
}
\description{
This function is used to get the GEO accession numbers of all studies in a platform.
}
