% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_functions.R
\name{insert_study_info}
\alias{insert_study_info}
\title{Insert study information}
\usage{
insert_study_info(pool, gds_id, disease, platform, organism, summary,
  title, pubmedid)
}
\arguments{
\item{disease}{Character string describing the disease or condition the 
study refers to}

\item{platform}{Character string specifying the GEO platform used by the study}

\item{organism}{Character string specifying the study organism.}

\item{study_id}{Character string with the GEO accession number to be added}

\item{keywords}{Character string with keywords describing the study to be added,
with multiple keywords separated by semicolons}
}
\description{
This function takes information about a study and adds it into the 
study_info table in the database.
}
